requirements.txt
Flask==2.2.2
gunicorn==20.1.0
torch==2.1.0
torchvision==0.15.0
Pillow==9.3.0
werkzeug==2.2.2
numpy==1.23.5
opencv-python==4.6.0.66
flask-cors==3.1.1
